* Renaming testFunctions.py to test_functions.py and updating test_* * Updating test classes to remove all reference to "names"
